+++ projects/journey-maps-client-testapp/src/app/app.component.spec.ts
@@ -1,20 +1,18 @@
 import {TestBed} from '@angular/core/testing';
 import {AppComponent} from './app.component';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
-import {JourneyMapsClientComponent} from '../../../journey-maps-client/src/lib/journey-maps-client.component';
-import {InfoBoxComponent} from '../../../journey-maps-client/src/lib/components/info-box/info-box.component';
 import {SourceHighlightComponent} from './components/source-highlight/source-highlight.component';
+import {JourneyMapsClientModule} from 'journey-maps-client';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
-        HttpClientTestingModule
+        HttpClientTestingModule,
+        JourneyMapsClientModule
       ],
       declarations: [
         AppComponent,
-        JourneyMapsClientComponent,
-        InfoBoxComponent,
         SourceHighlightComponent
       ],
     }).compileComponents();