+++ projects/journey-maps-client/src/lib/journey-maps-client.module.ts
@@ -2,29 +2,35 @@ import {NgModule} from '@angular/core';
 import {JourneyMapsClientComponent} from './journey-maps-client.component';
 import {HttpClientModule} from '@angular/common/http';
 import {CommonModule} from '@angular/common';
-import {InfoBoxComponent} from './components/info-box/info-box.component';
 import {SBB_ICON_REGISTRY_PROVIDER, SbbIconModule} from '@sbb-esta/angular-core';
-import {TextInfoBoxComponent} from './components/info-box/components/text-info-box/text-info-box.component';
-import {ButtonInfoBoxComponent} from './components/info-box/components/button-info-box/button-info-box.component';
-import {HtmlInfoBoxComponent} from './components/info-box/components/html-info-box/html-info-box.component';
 import {SbbAutocompleteModule, SbbButtonModule, SbbSearchModule} from '@sbb-esta/angular-public';
 import {SearchBarComponent} from './components/search-bar/search-bar.component';
 import {ReactiveFormsModule} from '@angular/forms';
 import {AngularResizedEventModule} from 'angular-resize-event';
 import {GlobalStylesComponent} from './components/global-styles/global-styles.component';
-import { LevelSwitchComponent } from './components/level-switch/level-switch.component';
+import {MarkerDetailsComponent} from './components/marker-details/marker-details.component';
+import {LevelSwitchComponent} from './components/level-switch/level-switch.component';
+import {TextInfoBlockComponent} from './components/marker-details/components/info-block-renderer/components/text-info-block/text-info-block.component';
+import {ButtonInfoBlockComponent} from './components/marker-details/components/info-block-renderer/components/button-info-block/button-info-block.component';
+import {HtmlInfoBlockComponent} from './components/marker-details/components/info-block-renderer/components/html-info-block/html-info-block.component';
+import {PopupComponent} from './components/marker-details/components/popup/popup.component';
+import {TeaserComponent} from './components/marker-details/components/teaser/teaser.component';
+import {InfoBlockRendererComponent} from './components/marker-details/components/info-block-renderer/info-block-renderer.component';
 
 
 @NgModule({
   declarations: [
     JourneyMapsClientComponent,
-    InfoBoxComponent,
-    TextInfoBoxComponent,
-    ButtonInfoBoxComponent,
-    HtmlInfoBoxComponent,
+    TeaserComponent,
+    TextInfoBlockComponent,
+    ButtonInfoBlockComponent,
+    HtmlInfoBlockComponent,
     SearchBarComponent,
     GlobalStylesComponent,
-    LevelSwitchComponent
+    LevelSwitchComponent,
+    PopupComponent,
+    MarkerDetailsComponent,
+    InfoBlockRendererComponent
   ],
   imports: [
     CommonModule,